Updated ReadMe. Changed RTRShader.cpp name. Fixed data type conversion warnings.
- Removed some white space in ShaderObject header.
- Updated ReadMe.
- Changed RTRShader.cpp name.
-Fixed data type conversion warnings.

+++ Src/Application.cpp
@@ -41,7 +41,7 @@ void Application::Run() {
         }
 
         currentTime = localCurrentTime;
-        timeDelta = localTimeDeltaMilliseconds;
+        timeDelta = (float)localTimeDeltaMilliseconds;
 
         //convert timeDelta to seconds
         float tempTimeDelta = timeDelta / 1000.0f;
